+++ system-config.js
@@ -17,14 +17,33 @@ System.config({
     }
 });
 
+// Regular expressions for Meteor package import syntax
+var appRegex = /^\{}\//;
+var packageRegex = /^{([\w-]*?):?([\w-]+)}/;
+var packageRegexBC = /^([\w-]+):([\w-]+)/;
+
 /**
- * Convert author:package/module to author_package/module
- * @param {string} name - unnormalized module name with colon
- * @returns {string} - unnormalized module name without unnecessary colon
+ * Convert Meteor package syntax to System.normalize friendly string.
+ * The `__author_package/foo` syntax in an internal implementation that is subject to change.
+ * You should never rely on it! Instead pass all your module names through System.normalizeSync
+ * @param {string} name - unnormalized module name with Meteor package syntax
+ * @returns {string} - unnormalized module name without Meteor package syntax
  */
 var normalizeMeteorPackageName = function (name) {
-    // @todo replace this with a regexp
-    return name.replace(':', '_');
+    name = name
+        .replace(appRegex, '') // {}/foo -> foo
+        .replace(packageRegex, '__$1_$2'); // {author:package}/foo -> __author_package/foo
+
+    if(packageRegexBC.test(name)){
+        // provide temporary backward compatibility for versions < 0.4 package syntax
+        console.warn([
+            '[Universe Modules]',
+            'You are using deprecated syntax for importing modules from a package.',
+            'Instead of', name, 'you should use', name.replace(packageRegexBC, '{$1:$2}')
+        ].join(' '));
+        return name.replace(packageRegexBC, '__$1_$2'); // author:package/foo -> __author_package/foo
+    }
+    return name;
 };
 
 /*
@@ -34,25 +53,27 @@ var normalizeMeteorPackageName = function (name) {
  */
 System.normalize = function (name, parentName, parentAddress) {
 
-    // Remove colon from package name
-    name = normalizeMeteorPackageName(name);
-
     // Allow foomodule.import syntax in import name (TypeScript support)
     if (name.slice(-7) === '.import') {
         name = name.slice(0, -7);
     }
 
     // Load original normalize
-    return _System.normalize.call(this, name, parentName, parentAddress);
+    return _System.normalize.call(this, normalizeMeteorPackageName(name), parentName, parentAddress)
+        .then(function (normalizedName) {
+            console.log('normalize', name, '->', normalizedName);
+            return normalizedName;
+        });
 };
 
 /*
  * name: the unnormalized module name
  * parentName: the canonical module name for the requesting module
- * parentAddress: the address of the requesting module
  */
-System.normalizeSync = function (name, parentName, parentAddress) {
-    return _System.normalizeSync.call(this, normalizeMeteorPackageName(name), parentName, parentAddress);
+System.normalizeSync = function (name, parentName) {
+    var normalizedName = _System.normalizeSync.call(this, normalizeMeteorPackageName(name), parentName);
+    console.log('normalizeSync', name, '->', normalizedName);
+    return normalizedName;
 };
 
 /*